[ticket5] almost there
- each (N-1)-port now gets its own pair of management queues reserved
  for incoming&outgoing CDAP messages
- support for multiple input queues per port
- all three QueueAlloc policies are now working everywhere except for
  the top layer in routers
- some renaming around RA so it's less confusing to others
- stopped using ModuleAccess

+++ src/DIF/RA/PDUForwardingTableEntry.cc
@@ -26,7 +26,8 @@ PDUForwardingTableEntry::PDUForwardingTableEntry()
 {
 }
 
-PDUForwardingTableEntry::PDUForwardingTableEntry(Address& destaddr, int qosid, RMTPort* port)
+PDUForwardingTableEntry::PDUForwardingTableEntry
+    (Address& destaddr, unsigned short qosid, RMTPort* port)
 :  destAddr(destaddr), qosId(qosid), rmtPort(port)
 {
 }
@@ -56,7 +57,7 @@ const Address& PDUForwardingTableEntry::getDestAddr() const
     return destAddr;
 }
 
-int PDUForwardingTableEntry::getQosId() const
+unsigned short PDUForwardingTableEntry::getQosId() const
 {
     return qosId;
 }
